requirements.txt
Django==1.7.1
Markdown==2.5.1
Pillow==2.6.1
PyYAML==3.11
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
celery==3.1.19
django-celery==3.1.17
django-cors-headers==0.13
django-extensions==1.4.6
django-filter==0.8
django-rest-swagger==0.3.4
django-suit==0.2.13
djangorestframework==2.4.4
etherpad-lite==0.2.1
hoerapi==1.3
ircutils3==0.2.2
kombu==3.0.29
modgrammar==0.10
nltk==3.1
numpy==1.10.1
osf==1.12
pamqp==1.6.1
psycopg2==2.5.4
python-dateutil==2.2
pytz==2015.7
rabbitpy==0.23.0
requests==2.5.1
rules==0.3
scikit-learn==0.17
scipy==0.16.1
six==1.10.0
